real frequency mesh: separate namepsace for grids
