+++ gf/test/mesh_test.cpp
@@ -6,13 +6,14 @@
 
 #include "gtest/gtest.h"
 #include "alps/gf/mesh.hpp"
+#include "alps/gf/grid.hpp"
 #include "gf_test.hpp"
 
 #include <boost/filesystem/operations.hpp>
 
 TEST(Mesh, RealFrequencyLoadSave) {
     namespace g=alps::gf;
-    g::mesh::linear_real_frequency_grid grid(-3,3,100);
+    g::grid::linear_real_frequency_grid grid(-3,3,100);
     g::real_frequency_mesh mesh1(grid);
     g::real_frequency_mesh mesh2;
     {
@@ -31,7 +32,7 @@ TEST(Mesh, RealFrequencyLoadSave) {
 TEST(Mesh, RealFrequencyMeshQuadric) {
     double spread = 5.0;
     int nfreq = 41;
-    alps::gf::mesh::quadratic_real_frequency_grid grid(spread, nfreq);
+    alps::gf::grid::quadratic_real_frequency_grid grid(spread, nfreq);
     alps::gf::real_frequency_mesh mesh1(grid);
     EXPECT_EQ(mesh1.extent(), nfreq);
 }
@@ -39,26 +40,18 @@ TEST(Mesh, RealFrequencyMeshQuadric) {
 TEST(Mesh, RealFrequencyMeshLogarithmic) {
     double tmax = 5, tmin = 0.001;
     int nfreq = 41;
-    alps::gf::mesh::logarithmic_real_frequency_grid grid(tmax, tmin, nfreq);
+    alps::gf::grid::logarithmic_real_frequency_grid grid(tmax, tmin, nfreq);
     alps::gf::real_frequency_mesh mesh1(grid);
     EXPECT_EQ(mesh1.extent(), nfreq);
-    double ratio1 = mesh1.points()[0]/mesh1.points()[1];
-    double ratio2 = mesh1.points()[2]/mesh1.points()[3];
-    EXPECT_NEAR(ratio1, ratio2, 1e-10);
 }
 
 TEST(Mesh, RealFrequencyMeshLinear) {
   double Emin = -5;
   double Emax = 5;
   int nfreq = 20;
-  alps::gf::mesh::linear_real_frequency_grid grid(Emin, Emax, nfreq);
+  alps::gf::grid::linear_real_frequency_grid grid(Emin, Emax, nfreq);
   alps::gf::real_frequency_mesh mesh1(grid);
   EXPECT_EQ(mesh1.extent(), nfreq);
-  double diff1 = mesh1.points()[0] - mesh1.points()[1];
-  double diff2 = mesh1.points()[2] - mesh1.points()[3];
-  EXPECT_NEAR(diff1, diff2, 1e-10);
-  EXPECT_NEAR(Emin, mesh1.points()[0], 1e-10);
-  EXPECT_NEAR(Emax, mesh1.points()[mesh1.points().size()-1], 1e-10);
 }
 
 TEST(Mesh, BosonicMatsubara) {
